+++ blocklets/core/api/src/routes/v1.ts
@@ -1,4 +1,4 @@
-import { getOpenAIV2 } from '@api/libs/ai-provider';
+import { getOpenAIV2, getReqModel } from '@api/libs/ai-provider';
 import {
   createRetryHandler,
   processChatCompletion,
@@ -45,7 +45,7 @@ router.post(
             type: 'chatCompletion',
             promptTokens: (usageData.usage?.inputTokens as number) || 0,
             completionTokens: (usageData.usage?.outputTokens as number) || 0,
-            model: req.body?.model as string,
+            model: getReqModel(req),
             modelParams: req.body?.options?.modelOptions,
           }).catch((err) => {
             logger.error('Create token usage v2 error', { error: err });
@@ -141,8 +141,8 @@ router.post(
       return '/v1/audio/transcriptions';
     },
     parseReqBody: false,
-    async proxyReqOptDecorator(proxyReqOpts) {
-      const { apiKey } = await getOpenAIV2();
+    async proxyReqOptDecorator(proxyReqOpts, srcReq) {
+      const { apiKey } = await getOpenAIV2(srcReq);
       proxyReqOpts.headers!.Authorization = `Bearer ${apiKey}`;
       return proxyReqOpts;
     },
@@ -158,8 +158,8 @@ router.post(
     proxyReqPathResolver() {
       return '/v1/audio/speech';
     },
-    async proxyReqOptDecorator(proxyReqOpts) {
-      const { apiKey } = await getOpenAIV2();
+    async proxyReqOptDecorator(proxyReqOpts, srcReq) {
+      const { apiKey } = await getOpenAIV2(srcReq);
       proxyReqOpts.headers!.Authorization = `Bearer ${apiKey}`;
       return proxyReqOpts;
     },